// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserType = {
  student: 'student',
  teacher: 'teacher'
} as const

export type UserType = (typeof UserType)[keyof typeof UserType]


export const Gender = {
  male: 'male',
  female: 'female',
  others: 'others'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const PaymentMethod = {
  wxPay: 'wxPay',
  stripe: 'stripe',
  balance: 'balance',
  cash: 'cash',
  payLater: 'payLater',
  payForMe: 'payForMe'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const NotificationType = {
  orderCreated: 'orderCreated',
  pickupReminder: 'pickupReminder',
  orderRefunded: 'orderRefunded',
  balanceToppedUp: 'balanceToppedUp',
  pointsEarned: 'pointsEarned',
  payLaterReminder: 'payLaterReminder'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const OrderStatus = {
  waiting: 'waiting',
  done: 'done'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const OrderType = {
  pickUp: 'pickUp',
  delivery: 'delivery'
} as const

export type OrderType = (typeof OrderType)[keyof typeof OrderType]


export const PaymentStatus = {
  notPaid: 'notPaid',
  paid: 'paid',
  refunded: 'refunded'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const UserAuditLogType = {
  login: 'login',
  blocked: 'blocked',
  unblocked: 'unblocked',
  permissionsUpdated: 'permissionsUpdated',
  balanceTransaction: 'balanceTransaction',
  balanceUsed: 'balanceUsed',
  pointsUpdated: 'pointsUpdated',
  orderCreated: 'orderCreated',
  orderSetStatus: 'orderSetStatus',
  orderPaymentSuccess: 'orderPaymentSuccess',
  orderPaymentFailed: 'orderPaymentFailed',
  orderRefunded: 'orderRefunded',
  couponUsed: 'couponUsed',
  upsertCategory: 'upsertCategory',
  upsertOptionType: 'upsertOptionType',
  upsertOptionItem: 'upsertOptionItem',
  upsertTag: 'upsertTag',
  upsertCouponCode: 'upsertCouponCode',
  upsertAd: 'upsertAd',
  upsertItemType: 'upsertItemType',
  deleteCategory: 'deleteCategory',
  deleteOptionType: 'deleteOptionType',
  deleteOptionItem: 'deleteOptionItem',
  deleteTag: 'deleteTag',
  deleteCouponCode: 'deleteCouponCode',
  deleteAd: 'deleteAd',
  deleteItemType: 'deleteItemType'
} as const

export type UserAuditLogType = (typeof UserAuditLogType)[keyof typeof UserAuditLogType]